backend/requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.39.1
crispy-bootstrap5==2024.2
cryptography==42.0.5
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
Django==4.2.11
django-appconf==1.0.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-cryptography==1.1
django-environ==0.11.2
django-health-check==3.18.1
django-import-export==3.3.8
django-jazzmin==2.6.1
django-phonenumber-field==7.3.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
et-xmlfile==1.1.0
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==21.2.0
idna==3.6
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
phonenumbers==8.13.34
pillow==10.3.0
promise==2.3
psutil==5.9.8
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python3-openid==3.2.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tablib==3.5.0
whitenoise
text-unidecode==1.3
typing-extensions==4.11.0
urllib3==2.2.1
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
